Merge Develop new

+++ mlonmcu/session/session.py
@@ -17,6 +17,7 @@
 # limitations under the License.
 #
 """Definition of a MLonMCU Run which represents a set of benchmarks in a session."""
+
 import os
 import shutil
 import tempfile
@@ -367,6 +368,7 @@ class Session:
             self.status = SessionStatus.CLOSED
         self.closed_at = datetime.now()
         self.session_lock.release()
-        os.remove(self.session_lock.lock_file)
+        if os.path.exists(self.session_lock.lock_file):
+            os.remove(self.session_lock.lock_file)
         if self.tempdir:
             self.tempdir.cleanup()
